# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: MyLocalHealth API
  version: 0.1.0
  summary: Local public-health context endpoints used by the MyLocalHealth app.
  description: >
    These endpoints support the MyLocalHealth dashboard. They aggregate public
    health, environmental, geocoding, news, and AI-assistant context. Responses
    are informational only and must not be interpreted as medical advice,
    diagnosis, or treatment.
servers:
  - url: https://mylocalhealth.org
    description: Production
  - url: http://localhost:3000
    description: Local development
paths:
...
  /api/risk:
    get:
      summary: Generate a compact risk snapshot for a ZIP code.
      description: >
        Validates the ZIP code, applies rate limiting, aggregates location,
        air quality, forecast, CDC respiratory signals, weather alerts, and
        health-equity context, then returns the transparent risk model output.
      parameters:
        - $ref: "#/components/parameters/ZipCode"
      responses:
        "200":
          description: Local risk snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskSnapshot"
        "400":
          $ref: "#/components/responses/ApiError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ApiError"
...
components:
  parameters:
    ZipCode:
      name: zipCode
      in: query
      required: true
      schema:
        type: string
        pattern: "^[0-9]{5}$"
      example: "80528"
...
  responses:
    ApiError:
      description: Request failed.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    RateLimited:
      description: Request was rate limited.
      headers:
        Retry-After:
          schema:
            type: string
          description: Seconds until the client should retry.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              retryAfterSeconds:
                type: integer
  schemas:
    RiskSnapshot:
      type: object
      properties:
        zipCode:
          type: string
        location:
          type: object
          additionalProperties: true
        generatedAt:
          type: string
          format: date-time
        risk:
          type: object
          properties:
            overall:
              type: string
            respiratory:
              type: string
            score:
              type: integer
            modelVersion:
              type: string
        signals:
          type: object
          additionalProperties: true
        forecast:
          type:
            - object
            - "null"
          additionalProperties: true
        equity:
          type:
            - object
            - "null"
          additionalProperties: true
        confidence:
          type: object
          additionalProperties: true
        featureSnapshot:
          type: object
          additionalProperties: true
        disclaimer:
          type: string
...